requirements.txt
pydantic
pydantic-ai
sympy
python-dotenv
latex2sympy2
transformers>=4.30.0
torch>=2.0.0
numpy>=1.21.0
tqdm>=4.65.0
streamlit